backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
httpx
sqlalchemy[asyncio]
asyncpg
alembic